Removed Home and Chat fragments and updated dependent code
